major: New features and moved files
Added a battery voltage reading function
Moved functions out of main.cpp to their own library (kinda messy)
Added Hardware files for different types of 3D models
Added preliminary BOM
Updated schematic
