--- Design Files/Electronics Design/desktop.ini
@@ -1,2 +1,0 @@
-[.ShellClassInfo]
-IconResource=C:\Program Files\Google\Drive File Stream\57.0.5.0\GoogleDriveFS.exe,23
