+++ Hardware/Remote Sensor 2.csv
@@ -1,0 +1,60 @@
+"Source:","C:\Users\nozol\Documents\5School\College Lv7\3rd Yr\SEMESTER 2\PROJ\PCB\Remote Sensor Rev 2\Remote Sensor 2.kicad_sch"
+"Date:","22/01/2023 17:43:59"
+"Tool:","Eeschema (6.0.7)"
+"Generator:","C:\Program Files\KiCad\6.0\bin\scripting\plugins/bom_csv_grouped_by_value.py"
+"Component Count:","29"
+
+"Individual Components:"
+
+"Item","Qty","Reference(s)","Value","LibPart","Footprint","Datasheet"
+"","","BT101","Battery_Cell","Device:Battery_Cell","Connector_PinHeader_2.54mm:PinHeader_1x02_P2.54mm_Vertical","~"
+"","","D101","LED","Remote-Sensor-2-rescue:LED_BGCR-Device","LED_THT:LED_D5.0mm-4_RGB_Wide_Pins",""
+"","","H101","MountingHole","Mechanical:MountingHole","MountingHole:MountingHole_5.5mm","~"
+"","","H102","MountingHole","Mechanical:MountingHole","MountingHole:MountingHole_5.5mm","~"
+"","","H103","MountingHole","Mechanical:MountingHole","MountingHole:MountingHole_5.5mm","~"
+"","","H104","MountingHole","Mechanical:MountingHole","MountingHole:MountingHole_5.5mm","~"
+"","","Q101","S8050","Transistor_BJT:S8050","Package_TO_SOT_THT:TO-92_Inline","http://www.unisonic.com.tw/datasheet/S8050.pdf"
+"","","R101","1k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P2.54mm_Vertical","~"
+"","","R102","1k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P2.54mm_Vertical","~"
+"","","R103","1k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P2.54mm_Vertical","~"
+"","","R104","1k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P10.16mm_Horizontal","~"
+"","","R105","1k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P10.16mm_Horizontal","~"
+"","","R106","2k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P10.16mm_Horizontal","~"
+"","","R107","2k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P10.16mm_Horizontal","~"
+"","","SC101","Solar_Cells","Device:Solar_Cells","Sensor_Footprints:Terminal_Block_bornier-2_P5.08_No_SS","~"
+"","","TP101","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"","","TP102","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"","","TP103","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"","","TP105","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"","","TP106","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"","","TP107","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"","","TP108","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"","","U101","CN3065_Breakout","Sensor-Parts:CN3065_Breakout","Sensor_Footprints:CN3065_Breakout",""
+"","","U102","Regulator_Breakout_3.3","Sensor-Parts:Regulator_Breakout_3.3","Connector_PinSocket_2.54mm:PinSocket_1x03_P2.54mm_Vertical",""
+"","","U103","RV_3028_Breakout","Sensor-Parts:RV_3028_Breakout","Connector_PinSocket_2.54mm:PinSocket_1x05_P2.54mm_Vertical",""
+"","","U104","RFM9x","Sensor-Parts:RFM9x","Sensor_Footprints:RF9x_Breakout",""
+"","","U105","Arduino_Pro_Mini","Remote-Sensor-2-rescue:Arduino_Pro_Mini-Arduino","Arduino:Arduino_Pro_Mini","https://www.sparkfun.com/products/11113"
+"","","U106","HDC1080_Breakout","Sensor-Parts:HDC1080_Breakout","Connector_PinSocket_2.54mm:PinSocket_1x04_P2.54mm_Vertical",""
+"","","U107","uSD_Card_Holder","Sensor-Parts:uSD_Card_Holder","Sensor_Footprints:PJRC_uSD",""
+
+
+
+"Collated Components:"
+
+"Item","Qty","Reference(s)","Value","LibPart","Footprint","Datasheet"
+"1","1","BT101","Battery_Cell","Device:Battery_Cell","Connector_PinHeader_2.54mm:PinHeader_1x02_P2.54mm_Vertical","~"
+"2","1","D101","LED","Remote-Sensor-2-rescue:LED_BGCR-Device","LED_THT:LED_D5.0mm-4_RGB_Wide_Pins",""
+"3","4","H101, H102, H103, H104","MountingHole","Mechanical:MountingHole","MountingHole:MountingHole_5.5mm","~"
+"4","1","Q101","S8050","Transistor_BJT:S8050","Package_TO_SOT_THT:TO-92_Inline","http://www.unisonic.com.tw/datasheet/S8050.pdf"
+"5","3","R101, R102, R103","1k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P2.54mm_Vertical","~"
+"6","2","R104, R105","1k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P10.16mm_Horizontal","~"
+"7","2","R106, R107","2k","Device:R","Resistor_THT:R_Axial_DIN0207_L6.3mm_D2.5mm_P10.16mm_Horizontal","~"
+"8","1","SC101","Solar_Cells","Device:Solar_Cells","Sensor_Footprints:Terminal_Block_bornier-2_P5.08_No_SS","~"
+"9","7","TP101, TP102, TP103, TP105, TP106, TP107, TP108","TestPoint","Connector:TestPoint","TestPoint:TestPoint_THTPad_D1.0mm_Drill0.5mm","~"
+"10","1","U101","CN3065_Breakout","Sensor-Parts:CN3065_Breakout","Sensor_Footprints:CN3065_Breakout",""
+"11","1","U102","Regulator_Breakout_3.3","Sensor-Parts:Regulator_Breakout_3.3","Connector_PinSocket_2.54mm:PinSocket_1x03_P2.54mm_Vertical",""
+"12","1","U103","RV_3028_Breakout","Sensor-Parts:RV_3028_Breakout","Connector_PinSocket_2.54mm:PinSocket_1x05_P2.54mm_Vertical",""
+"13","1","U104","RFM9x","Sensor-Parts:RFM9x","Sensor_Footprints:RF9x_Breakout",""
+"14","1","U105","Arduino_Pro_Mini","Remote-Sensor-2-rescue:Arduino_Pro_Mini-Arduino","Arduino:Arduino_Pro_Mini","https://www.sparkfun.com/products/11113"
+"15","1","U106","HDC1080_Breakout","Sensor-Parts:HDC1080_Breakout","Connector_PinSocket_2.54mm:PinSocket_1x04_P2.54mm_Vertical",""
+"16","1","U107","uSD_Card_Holder","Sensor-Parts:uSD_Card_Holder","Sensor_Footprints:PJRC_uSD",""
