+++ lib/lmicJobs/internaldefs.h
@@ -1,0 +1,23 @@
+// Duplicate of SensorNodeDefs.h because I don't know what I'm doing
+#ifndef UGO_INT_DEF
+#define UGO_INT_DEF
+
+// Pin definitions
+#define LORA_SS     10
+#define LORA_RST    17
+#define LORA_PIN0   14
+#define LORA_PIN1   5
+
+#define LED_R       7
+#define LED_G       8
+#define LED_B       9
+
+#define BATT_SENSE  16
+#define BATT_CTRL   6
+
+#define RTC_INT     3
+
+// Other
+#define WAKE_INTERVAL   600  // Unit: Seconds
+
+#endif // UGO_INT_DEF
