+++ platformio.ini
@@ -9,27 +9,27 @@
 ; https://docs.platformio.org/page/projectconf.html
 
 [platformio]
-default_envs = promini
-
-[env:nanoatmega328new]
-platform = atmelavr
-board = nanoatmega328new
-framework = arduino
-monitor_speed = 115200
-build_type = debug
-lib_deps = 
-	mcci-catena/MCCI LoRaWAN LMIC library @ ^4.1.1
-	rocketscream/Low-Power@^1.81
-	closedcube/ClosedCube HDC1080@^1.3.2
-	macroyau/RV-3028-C7@^1.1.2
+default_envs = pro8MHzatmega328
+description = My final year project for ECE
 
 [env:promini]
 platform = atmelavr
+board = pro16MHzatmega328
+framework = arduino
+lib_deps = 
+	mcci-catena/MCCI LoRaWAN LMIC library@^4.1.1
+	lowpowerlab/LowPower_LowPowerLab@^2.2
+	closedcube/ClosedCube HDC1080@^1.3.2
+	macroyau/RV-3028-C7@^1.1.2
+monitor_speed = 115200
+
+[env:pro8MHzatmega328]
+platform = atmelavr
 board = pro8MHzatmega328
 framework = arduino
-monitor_speed = 115200
 lib_deps = 
-	rocketscream/Low-Power@^1.81
+	mcci-catena/MCCI LoRaWAN LMIC library@^4.1.1
+	lowpowerlab/LowPower_LowPowerLab@^2.2
 	closedcube/ClosedCube HDC1080@^1.3.2
-	mcci-catena/MCCI LoRaWAN LMIC library @ ^4.1.1
 	macroyau/RV-3028-C7@^1.1.2
+monitor_speed = 115200
