+++ src/main.cpp
@@ -1,37 +1,11 @@
 #include "Arduino.h"
 #include <lmic.h>
 #include <hal/hal.h>
-#include <SPI.h>
-#include <Wire.h>
-#include <LowPower.h>
-#include <ClosedCube_HDC1080.h>
-#include <RV3028C7.h>
+#include <SensorNodeDefs.h>
+#include <lmicJobs.h>
 
-// Global Variables
-RV3028C7 rtc;
-ClosedCube_HDC1080 hdc1080;
-uint8_t LED_PIN = 8;
-static uint8_t mydata[4];
-
-// Function Declarations
-void onEvent (ev_t ev);
-void initFunc(osjob_t* j);
-void wakeUp(osjob_t* j);
-void readSensor(osjob_t* j);
-void logData(osjob_t* j);
-void do_send(osjob_t* j);
-void sleep(osjob_t* j);
-void setupChannelsEU868();
-
-
-
-// Job structs
-static osjob_t initJob;
-static osjob_t wakeJob;
-static osjob_t readJob;
-static osjob_t logJob;
-static osjob_t sendjob;
-static osjob_t sleepJob;
+// Uncommonet to transmit on only 868.1 MHz THIS VIOLATES LORAWAN/TTN REGULATIONS
+#define INDOOR_TESTING
 
 // LoRaWAN NwkSKey, MSB first
 static const PROGMEM u1_t NWKSKEY[16] = { 0x00, 0xC8, 0x85, 0x1A, 0xB2, 0x03, 0x36, 0xD5, 0x01, 0x44, 0x5F, 0x79, 0xC3, 0x9C, 0xF2, 0x4E };
@@ -42,20 +16,13 @@ static const u1_t PROGMEM APPSKEY[16] = { 0x59, 0x45, 0x55, 0xE0, 0xA6, 0x40, 0x
 // LoRaWAN end-device address MSB first
 static const u4_t DEVADDR = 0x260B540A ; // <-- Change this address for every node!
 
-const unsigned TX_INTERVAL = 600; // 10 minuites
-
 // Pin mapping
 const lmic_pinmap lmic_pins = {
-    .nss = 10,                       // chip select on feather (rf95module) CS
+    .nss = LORA_SS,
     .rxtx = LMIC_UNUSED_PIN,
-    .rst = 9,                       // reset pin
-    .dio = {15, 6, LMIC_UNUSED_PIN}, 
+    .rst = LORA_RST,
+    .dio = {LORA_PIN0, LORA_PIN1, LMIC_UNUSED_PIN},
 };
-
-void wakeUp()
-{
-    // Just a handler for the pin interrupt.
-}
 
 void setup() {
     CLKPR = 0x80; // (1000 0000) enable change in clock frequency
@@ -64,11 +31,8 @@ void setup() {
 
     Serial.begin(115200);
     Serial.println(F("Starting Sensor Node"));
-    pinMode(LED_PIN, OUTPUT);
-    Wire.begin();
-    rtc.begin();
-    hdc1080.begin(0x40);
-    
+    pinMode(LED_G, OUTPUT);
+    pinMode(BATT_CTRL, OUTPUT);
 
     // LMIC init
     os_init();
@@ -101,202 +65,10 @@ void setup() {
     // Set data rate and transmit power for uplink
     LMIC_setDrTxpow(DR_SF7,20);
     
+    // Program flow starts here. Defined in **TBD**
     initFunc(&initJob);
 }
 
 void loop() {
     os_runloop_once();
-}
-
-void initFunc(osjob_t* j){
-    // Blink on startup
-    for (int i = 0; i < 3; i++) {
-        delay(500);
-        digitalWrite(8, HIGH);
-        delay(100);
-        digitalWrite(8, LOW);
-        delay(100);
-    }
-    Serial.println("Init Job");
-    Serial.flush();
-
-    os_setCallback(&readJob, readSensor);
-}
-
-void wakeUp(osjob_t* j){
-    Serial.println("Wake");
-    Serial.flush();
-}
-
-void readSensor(osjob_t* j){
-    Serial.println("Reading sensor");
-
-    // Compacting data for transmission
-    int senseVal = hdc1080.readHumidity() * 100;
-    mydata[0] = highByte(senseVal);
-    mydata[1] = lowByte(senseVal);
-    
-    senseVal = hdc1080.readTemperature() * 100;
-    mydata[2] = highByte(senseVal);
-    mydata[3] = lowByte(senseVal);
-    os_setCallback(&logJob, logData);
-}
-
-void logData(osjob_t* j){
-    // SD Card not yet implemented
-    Serial.println("Logging");
-    Serial.flush();
-    os_setCallback(&sendjob, do_send);
-    ////os_setTimedCallback(&sendjob, os_getTime()+sec2osticks(TX_INTERVAL), do_send);
-}
-
-//I'm still not sure what the job argument is for
-void do_send(osjob_t* j){ // The job struct is passed to make sure that the cb calls the correct funtion for this job
-    Serial.println("Sending");
-    Serial.flush();
-    // Check if there is not a current TX/RX job running
-    if (LMIC.opmode & OP_TXRXPEND) {
-        Serial.println(F("OP_TXRXPEND, not sending"));
-    } else {
-        // Read sensors
-        // Log Data 
-        // Prepare upstream data transmission at the next possible time.
-        LMIC_setTxData2(1, mydata, sizeof(mydata), 0);
-        Serial.println(F("Packet queued"));
-    }
-    // Next TX is scheduled after TX_COMPLETE event.
-}
-
-void sleep(osjob_t* j){
-    // Enable sleep timer interrupt and go to sleep
-    
-    attachInterrupt(digitalPinToInterrupt(2), wakeUp, FALLING);
-    Serial.println("Going to sleep...");
-    Serial.flush();
-    rtc.setPeriodicCountdownTimer(TX_INTERVAL, TIMER_1HZ);
-    rtc.enableInterrupt(INTERRUPT_PERIODIC_COUNTDOWN_TIMER);
-    rtc.startPeriodicCountdownTimer();
-    LowPower.powerDown(SLEEP_FOREVER, ADC_OFF, BOD_OFF);
-
-    Serial.println("Waking up!!!");
-    Serial.flush();
-    rtc.stopPeriodicCountdownTimer();
-    rtc.clearInterrupt(INTERRUPT_PERIODIC_COUNTDOWN_TIMER);
-    rtc.disableInterrupt(INTERRUPT_PERIODIC_COUNTDOWN_TIMER);
-    detachInterrupt(digitalPinToInterrupt(2));
-    
-    os_setCallback(&readJob, readSensor);
-}
-
-void setupChannelsEU868(){    
-   LMIC_setupChannel(0, 868100000, DR_RANGE_MAP(DR_SF12, DR_SF7),  BAND_CENTI);      // g-band
-
-    LMIC_disableChannel(1);
-    LMIC_disableChannel(2);
-    LMIC_disableChannel(3);
-    LMIC_disableChannel(4);
-    LMIC_disableChannel(5);
-    LMIC_disableChannel(6);
-    LMIC_disableChannel(7);
-    LMIC_disableChannel(8); 
-}
-
-void onEvent (ev_t ev) {
-    Serial.print(os_getTime());
-    Serial.print(": ");
-    switch(ev) {
-        /*case EV_SCAN_TIMEOUT:
-            Serial.println(F("EV_SCAN_TIMEOUT"));
-            break;
-        case EV_BEACON_FOUND:
-            Serial.println(F("EV_BEACON_FOUND"));
-            break;
-        case EV_BEACON_MISSED:
-            Serial.println(F("EV_BEACON_MISSED"));
-            break;
-        case EV_BEACON_TRACKED:
-            Serial.println(F("EV_BEACON_TRACKED"));
-            break;            
-        case EV_LOST_TSYNC:
-            Serial.println(F("EV_LOST_TSYNC"));
-            break;
-        case EV_JOINING:
-            Serial.println(F("EV_JOINING"));
-            break;
-        case EV_JOINED:
-            Serial.println(F("EV_JOINED"));
-            break;
-        
-        || This event is defined but not used in the code. No
-        || point in wasting codespace on it.
-        ||
-        || case EV_RFU1:
-        ||     Serial.println(F("EV_RFU1"));
-        ||     break;
-        
-        case EV_JOIN_FAILED:
-            Serial.println(F("EV_JOIN_FAILED"));
-            break;
-        case EV_REJOIN_FAILED:
-            Serial.println(F("EV_REJOIN_FAILED"));
-            break;
-        case EV_JOIN_TXCOMPLETE:
-            Serial.println(F("EV_JOIN_TXCOMPLETE: no JoinAccept"));
-            break;*/
-        case EV_TXCOMPLETE:
-            //sleep emits tx start
-            Serial.println(F("EV_TXCOMPLETE (includes waiting for RX windows)"));
-            if (LMIC.txrxFlags & TXRX_ACK)
-              Serial.println(F("Received ack"));
-            if (LMIC.dataLen) {
-              Serial.println(F("Received "));
-              Serial.println(LMIC.dataLen);
-              Serial.println(F(" bytes of payload"));
-            }
-            
-            // Schedule next transmission
-            ////os_setTimedCallback(&sendjob, os_getTime()+sec2osticks(TX_INTERVAL), do_send);  
-            digitalWrite(LED_PIN, HIGH);
-            delay(200);
-            digitalWrite(LED_PIN, LOW);
-            os_setCallback(&sleepJob, sleep);
-            
-            break;
-        case EV_RESET:
-            Serial.println(F("EV_RESET"));
-            break;
-        case EV_RXCOMPLETE:
-            // data received in ping slot
-            Serial.println(F("EV_RXCOMPLETE"));
-            break;
-        case EV_LINK_DEAD:
-            Serial.println(F("EV_LINK_DEAD"));
-            break;
-        case EV_LINK_ALIVE:
-            Serial.println(F("EV_LINK_ALIVE"));
-            break;
-        /*
-        || This event is defined but not used in the code. No
-        || point in wasting codespace on it.
-        ||
-        || case EV_SCAN_FOUND:
-        ||    Serial.println(F("EV_SCAN_FOUND"));
-        ||    break;
-        */
-        case EV_TXSTART:
-            Serial.println(F("EV_TXSTART"));
-            //readsensor
-            //transmitsata emits txcomplete
-            break;
-        case EV_TXCANCELED:
-            Serial.println(F("EV_TXCANCELED"));
-            break;
-        case EV_RXSTART:
-            /* do not print anything -- it wrecks timing */
-            break;
-        default:
-            Serial.print(F("Unknown event: "));
-            Serial.println((unsigned) ev);
-            break;
-    }
 }
